ref(shared): move interop Wrapper read model into Shared
InteropFacadeInterface (in Shared) referenced Interop\Domain\ReadModel\Wrapper,
creating a Shared->Interop dependency cycle. The DTO has no dependencies and
now lives in Shared\Interop.
